requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.7.4
exceptiongroup==1.2.2
ffmpeg-python==0.2.0
frozenlist==1.4.1
future==1.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imouapi==1.0.13
mkdocs-material-extensions==1.3.1
multidict==6.0.5
python-telegram-bot==21.4
sniffio==1.3.1
typing_extensions==4.12.2
yarl==1.9.4
